fix: errores en consultas personalizadas en ContractRepository, refactor: implementación format() para mensajes parametrizados y uso de lombok en inyección de dependencias, feat: implementación de clase y enums para mensajes centralizados

+++ backend/administrative-system-backend/src/main/java/com/gg/administrative_system_backend/contract/Contract.java
@@ -20,5 +20,5 @@ public class Contract {
         @Builder.Default
         private boolean status = true;
         @Builder.Default
-        private float totalExpenses = 0;
+        private double totalExpenses = 0.0;
 }
